build our own version of optipng - the debian one is old and buggy. use tmp dir for optimizing so i can run multiple in parralel (derp)

+++ build/optimize.sh
@@ -1,35 +1,46 @@
 #!/bin/bash
 
+TMP=$(mktemp -d /tmp/optimize-XXXXXXXX)
+if [ ! -d $TMP ]; then
+	echo "Failed to create temp dir"
+	exit 1
+fi
+TEMP1="$TMP/stage1.png"
+TEMP2="$TMP/stage2.png"
+
 for f in $*
 do
 
-	echo "Processing $f";
+	rm -f "$TEMP1" "$TEMP2"
 
-	rm -f /tmp/stage1.png /tmp/stage2.png
+	echo "Processing $f";
 
 	ORIG_SIZE=$(stat -c%s "$f")
 	echo "  start    $ORIG_SIZE"
 
-	pngcrush -brute -reduce -s $f /tmp/stage1.png
-	SIZE2=$(stat -c%s "/tmp/stage1.png")
+	pngcrush -brute -reduce -noforce -s $f $TEMP1
+	if [ ! -f $TEMP1 ]; then
+		cp $f $TEMP1
+	fi
+	SIZE2=$(stat -c%s "$TEMP1")
 	echo "  pngcrush $SIZE2"
 
-	optipng -o7 /tmp/stage1.png > /dev/null
-	SIZE3=$(stat -c%s "/tmp/stage1.png")
+	/usr/local/bin/optipng -o7 -zm1-9 -quiet $TEMP1
+	SIZE3=$(stat -c%s "$TEMP1")
 	echo "  optipng  $SIZE3"
 
-	pngout-static -q /tmp/stage1.png /tmp/stage2.png
-	SIZE4=$(stat -c%s "/tmp/stage2.png")
+	pngout-static -q $TEMP1 $TEMP2
+	SIZE4=$(stat -c%s $TEMP2)
 	echo "  pngout   $SIZE4"
 
-	advpng --recompress --shrink-insane --quiet /tmp/stage2.png
-	SIZE5=$(stat -c%s "/tmp/stage2.png")
+	advpng --recompress --shrink-insane --quiet $TEMP2
+	SIZE5=$(stat -c%s $TEMP2)
 	echo "  advpng   $SIZE5"
 
-	node ~/zopfli-png/zopfli-png.js --i1000 --silent /tmp/stage2.png
-	SIZE6=$(stat -c%s "/tmp/stage2.png")
+	node ~/zopfli-png/zopfli-png.js --i1000 --silent $TEMP2
+	SIZE6=$(stat -c%s $TEMP2)
 	echo "  zopfli   $SIZE6"
 
 	rm -f "$f"
-	cp /tmp/stage2.png "$f"
+	cp $TEMP2 "$f"
 done
